Refactor distance.rs and motor.rs to be private
Their struct impls have also been modified to take Pins directly in
their constructors and no longer export them, as they are meant to be
constructed by another struct which will export the pins for them. This
will also reduce the amount of time required to sleep while the pins are
exported.