Use futures for movement and sensing
This change also requires removing the Drop implementations from
distance::Sensor and motor::Controller, moving the functionality to
a new unexport method. controller::Controller has gained a drop
implementation that calls these methods. This is done to ensure that
the cloned motors and sensors do not invalidate the pins too early.